Moved FO pages to classic directory
