+++ src/pages/FO/classic/category/index.ts
@@ -7,9 +7,9 @@ const psVersion = testContext.getPSVersion();
 /* eslint-disable global-require */
 function requirePage(): FoCategoryPageInterface {
   if (semver.gte(psVersion, '8.0.0')) {
-    return require('@versions/8.0.0/pages/FO/category');
+    return require('@versions/8.0.0/pages/FO/classic/category');
   }
-  return require('@versions/8.0.0/pages/FO/category');
+  return require('@versions/8.0.0/pages/FO/classic/category');
 }
 /* eslint-enable global-require */
