+++ src/pages/FO/classic/login/index.ts
@@ -7,9 +7,9 @@ const psVersion = testContext.getPSVersion();
 /* eslint-disable global-require, @typescript-eslint/no-var-requires */
 function requirePage(): FoLoginPageInterface {
   if (semver.gte(psVersion, '8.0.0')) {
-    return require('@versions/8.0.0/pages/FO/login').loginPage;
+    return require('@versions/8.0.0/pages/FO/classic/login').loginPage;
   }
-  return require('@versions/8.0.0/pages/FO/login').loginPage;
+  return require('@versions/8.0.0/pages/FO/classic/login').loginPage;
 }
 /* eslint-enable global-require, @typescript-eslint/no-var-requires */
